Move code to load thingpedia primitive templates into the construct templates
The code depends heavily on the specific templates used, in the
sense that it depends on the names of the non-terminals to use
for primitive templates, constants and output parameters.

Also, it is very much thingtalk specific - other languages don't
necessarily do the same things for input/output parameters.
