+++ languages/thingtalk/common-constants.genie
