+++ languages/thingtalk/en/parameters.genie
@@ -17,9 +17,10 @@ const Type = ThingTalk.Type;
 
 // import the combinator library
 const C = require('../ast_manip');
-}
 
-import '../common-parameters';
+// import thingpedia info
+const _tpLoader = require('../load-thingpedia');
+}
 
 // these are autogenerated and depends on projection_*, which is also
 // autogenerated
@@ -51,7 +52,7 @@ single_projection_String = table:complete_table => {
         if (pname === 'picture_url')
             return null;
         let ptype = table.schema.out[pname];
-        if ($options.types.id.has(C.typeToStringSafe(ptype)))
+        if (_tpLoader.types.id.has(C.typeToStringSafe(ptype)))
             return null;
     }
     return new Ast.Table.Projection(table, ['$event'], table.schema);
@@ -67,7 +68,7 @@ single_stream_projection_String = 'new' table:complete_table => {
         if (pname === 'picture_url')
             return null;
         let ptype = table.schema.out[pname];
-        if ($options.types.id.has(C.typeToStringSafe(ptype)))
+        if (_tpLoader.types.id.has(C.typeToStringSafe(ptype)))
             return null;
     }
     return new Ast.Stream.Projection(new Ast.Stream.Monitor(table, null, table.schema), ['$event'], table.schema);
@@ -102,7 +103,7 @@ the_out_param_Array__Any = {
     }
 }
 
-for (let [typestr, type] of $options.types.all) {
+for (let [typestr, type] of _tpLoader.types.all) {
     // don't access booleans or enums out arguments generically, as that rarely makes sense
     // (and when it does, you probably want a macro and maybe and edge trigger)
     if (!type.isEnum && !type.isBoolean) {
@@ -115,13 +116,13 @@ for (let [typestr, type] of $options.types.all) {
             }
         }
 
-        if (!$options.types.id.has(typestr)) {
-            if (!$options.flags.projection_with_filter) {
+        if (!_tpLoader.types.id.has(typestr)) {
+            if (!_tpLoader.flags.projection_with_filter) {
                 $('projection_' + typestr) = outParam:$('the_out_param_' + typestr) 'of' table:complete_table => {
                     const name = outParam.name;
                     if (!table.schema.out[name] || !Type.isAssignable(table.schema.out[name], type))
                         return null;
-                    if (name === 'picture_url' && $options.flags.turking)
+                    if (name === 'picture_url' && _tpLoader.flags.turking)
                         return null;
                     const newSchema = table.schema.filterArguments((arg) => arg.direction !== Ast.ArgDirection.OUT || arg.name === name);
                     return new Ast.Table.Projection(table, [name], newSchema);
@@ -141,7 +142,7 @@ for (let [typestr, type] of $options.types.all) {
                     const name = outParam.name;
                     if (!table.schema.out[name] || !Type.isAssignable(table.schema.out[name], type))
                         return null;
-                    if (name === 'picture_url' && $options.flags.turking)
+                    if (name === 'picture_url' && _tpLoader.flags.turking)
                         return null;
                     const newSchema = table.schema.filterArguments((arg) => arg.direction !== Ast.ArgDirection.OUT || arg.name === name);
                     return new Ast.Table.Projection(table, [name], newSchema);
@@ -161,8 +162,8 @@ for (let [typestr, type] of $options.types.all) {
             projection_Any = $('projection_' + typestr);
         }
 
-        if ($options.types.id.has(typestr)) {
-            if (!$options.flags.projection_with_filter) {
+        if (_tpLoader.types.id.has(typestr)) {
+            if (!_tpLoader.flags.projection_with_filter) {
                 $('single_projection_' + typestr) = table:complete_table => {
                     for (let pname in table.schema.out) {
                         if (table.schema.out[pname].equals(type))
@@ -307,7 +308,7 @@ placeholder_Entity__tt__url = {
     'some link' => Ast.Value.Undefined(true);
 }
 
-for (let [pname, [typestr, pcanonical]] of $options.params.in.values()) {
+for (let [pname, [typestr, pcanonical]] of _tpLoader.params.in.values()) {
     if (typestr.startsWith('Enum__')) {
         // replace $-placeholders with "some X" where X is based on parameter name
         if (pname !== 'power' && pname !== 'state' && pname !== 'p_power' && pname !== 'p_state') {
@@ -331,7 +332,7 @@ for (let [pname, [typestr, pcanonical]] of $options.params.in.values()) {
     // don't parameter pass booleans or enums, as that rarely makes sense
     if (typestr !== 'Boolean' && !typestr.startsWith('Enum__')) {
         if (pname !== 'count' && pname !== 'p_count') {
-            if ($options.flags.undefined_filter) {
+            if (_tpLoader.flags.undefined_filter) {
                 // replace $-placeholders with "some X" where X is based on parameter name
                 thingpedia_query = q:thingpedia_query #('some ' + pcanonical) [-> pname { isUndefined: true }] =>
                     C.replacePlaceholderWithUndefined(q, pname, typestr);
@@ -356,7 +357,7 @@ for (let [pname, [typestr, pcanonical]] of $options.params.in.values()) {
         action_replace_param_with_table = a:thingpedia_action p:$('single_projection_' + typestr) [-> pname] => C.actionReplaceParamWithTable(a, pname, p);
         action_replace_param_with_stream = a:thingpedia_action p:$('single_stream_projection_' + typestr) [-> pname] => C.actionReplaceParamWithStream(a, pname, p);
 
-        if ($options.types.id.has(typestr)) {
+        if (_tpLoader.types.id.has(typestr)) {
             forward_get_do_command = command:forward_get_do_command ('it' | 'that' | 'them') [-> pname] => {
                 for (let joinArg in command.table.schema.out) {
                     if (typestr === C.typeToStringSafe(command.table.schema.out[joinArg]))
@@ -404,7 +405,7 @@ for (let [pname, [typestr, pcanonical]] of $options.params.in.values()) {
                     if (pname === 'picture_url')
                         return null;
                     let ptype = command.table.schema.out[pname];
-                    if ($options.types.id.has(C.typeToStringSafe(ptype)))
+                    if (_tpLoader.types.id.has(C.typeToStringSafe(ptype)))
                         return null;
                 }
                 return C.getDoCommand(command, pname, new Ast.Value.Event(null));
@@ -415,7 +416,7 @@ for (let [pname, [typestr, pcanonical]] of $options.params.in.values()) {
                     if (pname === 'picture_url')
                         return null;
                     let ptype = rule.stream.schema.out[pname];
-                    if ($options.types.id.has(C.typeToStringSafe(ptype)))
+                    if (_tpLoader.types.id.has(C.typeToStringSafe(ptype)))
                         return null;
                 }
                 return C.whenDoRule(rule, pname, new Ast.Value.Event(null));
@@ -426,7 +427,7 @@ for (let [pname, [typestr, pcanonical]] of $options.params.in.values()) {
                     if (pname === 'picture_url')
                         return null;
                     let ptype = stream.stream.schema.out[pname];
-                    if ($options.types.id.has(C.typeToStringSafe(ptype)))
+                    if (_tpLoader.types.id.has(C.typeToStringSafe(ptype)))
                         return null;
                 }
                 return C.whenGetStream(stream, pname, new Ast.Value.Event(null));
