+++ languages/thingtalk/en/stream_tables.genie
@@ -18,6 +18,8 @@ const Type = ThingTalk.Type;
 // import the combinator library
 const C = require('../ast_manip');
 
+// import thingpedia info
+const _tpLoader = require('../load-thingpedia');
 }
 
 
@@ -38,14 +40,14 @@ one_if_filter_table = {
     !nofilter table:complete_table 'if' filter:atom_filter => {
         if (!C.checkFilter(table, filter))
             return null;
-        return C.addFilter(table, filter, $options);
+        return C.addFilter(table, filter, _tpLoader);
     };
 }
 two_if_filter_table = {
     !nofilter table:one_if_filter_table 'and' filter:atom_filter => {
         if (!C.checkFilter(table, filter))
             return null;
-        return C.addFilter(table, filter, $options);
+        return C.addFilter(table, filter, _tpLoader);
     };
 }
 with_filtered_table = {
@@ -60,7 +62,7 @@ one_with_filter_table = {
             return null;
         if (!C.checkFilter(table, filter))
             return null;
-        return C.addFilter(table, filter, $options);
+        return C.addFilter(table, filter, _tpLoader);
     };
 }
 two_with_filter_table = {
@@ -69,7 +71,7 @@ two_with_filter_table = {
             return null;
         if (!C.checkFilter(table, filter))
             return null;
-        return C.addFilter(table, filter, $options);
+        return C.addFilter(table, filter, _tpLoader);
     };
 }
 
@@ -78,7 +80,7 @@ npi_filtered_table = {
     !nofilter table:complete_table ('that' | 'which') 'is the' filter:npi_filter => {
         if (!C.checkFilter(table, filter))
             return null;
-        return C.addFilter(table, filter, $options);
+        return C.addFilter(table, filter, _tpLoader);
     };
 }
 
@@ -86,7 +88,7 @@ pvp_filtered_table = {
     !nofilter table:complete_table filter:pvp_filter => {
         if (!C.checkFilter(table, filter))
             return null;
-        return C.addFilter(table, filter, $options);
+        return C.addFilter(table, filter, _tpLoader);
     };
 }
 
@@ -94,7 +96,7 @@ avp_filtered_table = {
     !nofilter table:complete_table ('that' | 'which' | '') filter:avp_filter => {
         if (!C.checkFilter(table, filter))
             return null;
-        return C.addFilter(table, filter, $options);
+        return C.addFilter(table, filter, _tpLoader);
     };
 }
 
@@ -102,7 +104,7 @@ npv_filtered_table = {
     !nofilter table:complete_table ('that' | 'which') 'is a' filter:npv_filter => {
         if (!C.checkFilter(table, filter))
             return null;
-        return C.addFilter(table, filter, $options);
+        return C.addFilter(table, filter, _tpLoader);
     };
 }
 
@@ -110,13 +112,13 @@ apv_filtered_table = {
     !nofilter table:complete_table ('that' | 'which') 'is' filter:apv_filter => {
         if (!C.checkFilter(table, filter))
             return null;
-        return C.addFilter(table, filter, $options);
+        return C.addFilter(table, filter, _tpLoader);
     };
 
     !nofilter filter:apv_filter table:complete_table => {
         if (!C.checkFilter(table, filter))
             return null;
-        return C.addFilter(table, filter, $options);
+        return C.addFilter(table, filter, _tpLoader);
     };
 }
 
@@ -306,12 +308,12 @@ arg_min_max_table = {
 edge_stream = {
     !turking ('when' | 'if') 'the' p:projection_Any ('becomes' | 'becomes equal to') x:constant_Any => {
          if (x.getType().isString)
-             return C.makeEdgeFilterStream(p, '=~', x, $options);
+             return C.makeEdgeFilterStream(p, '=~', x, _tpLoader);
          else
-             return C.makeEdgeFilterStream(p, '==', x, $options);
+             return C.makeEdgeFilterStream(p, '==', x, _tpLoader);
     };
-    ('when' | 'if') 'the' p:projection_Numeric ('becomes greater than' | 'becomes higher than' | 'goes above' | 'increases above') x:constant_Numeric => C.makeEdgeFilterStream(p, '>=', x, $options);
-    ('when' | 'if') 'the' p:projection_Numeric ('becomes smaller than' | 'becomes lower than' | 'goes below' | 'decreases below') x:constant_Numeric => C.makeEdgeFilterStream(p, '<=', x, $options);
+    ('when' | 'if') 'the' p:projection_Numeric ('becomes greater than' | 'becomes higher than' | 'goes above' | 'increases above') x:constant_Numeric => C.makeEdgeFilterStream(p, '>=', x, _tpLoader);
+    ('when' | 'if') 'the' p:projection_Numeric ('becomes smaller than' | 'becomes lower than' | 'goes below' | 'decreases below') x:constant_Numeric => C.makeEdgeFilterStream(p, '<=', x, _tpLoader);
 }
 
 stream = {
@@ -331,7 +333,7 @@ stream = {
                 return null;
             let outParams = Object.keys(proj.table.schema.out);
             let stream;
-            if (outParams.length === 1 && $options.flags.turking)
+            if (outParams.length === 1 && _tpLoader.flags.turking)
                 return null;
             if (outParams.length === 1)
                 stream = C.tableToStream(proj.table, null);
@@ -343,7 +345,7 @@ stream = {
     !nofilter ('when' | 'if' | 'in case' | 'whenever' | 'any time' | 'should' | 'anytime') table:complete_table 'change and' filter:edge_filter => {
         if (!table.schema.is_monitorable || !C.checkFilter(table, filter) || table.schema.is_list)
             return null;
-        table = C.addFilter(table, filter, $options);
+        table = C.addFilter(table, filter, _tpLoader);
         if (!table)
             return null;
         return C.tableToStream(table, null);
@@ -351,9 +353,9 @@ stream = {
     !nofilter ('when' | 'if' | 'in case' | 'whenever' | 'any time' | 'should' | 'anytime') table:complete_table 'change and' filter:atom_filter => {
         if (!table.schema.is_monitorable || !C.checkFilter(table, filter))
             return null;
-        if ($options.flags.turking && table.schema.is_list)
+        if (_tpLoader.flags.turking && table.schema.is_list)
             return null;
-        table = C.addFilter(table, filter, $options);
+        table = C.addFilter(table, filter, _tpLoader);
         if (!table)
             return null;
         return C.tableToStream(table, null);
