+++ languages/thingtalk/zh-cn/aggregation.genie
