+++ languages/thingtalk/zh-cn/constants.genie
