+++ languages/thingtalk/zh-cn/parameters.genie
@@ -17,6 +17,9 @@ const Type = ThingTalk.Type;
 
 // import the combinator library
 const C = require('../ast_manip');
+
+// import thingpedia info
+const _tpLoader = require('../load-thingpedia');
 }
 
 import '../common-parameters';
@@ -51,7 +54,7 @@ single_projection_String = table:complete_table => {
         if (pname === 'picture_url')
             return null;
         let ptype = table.schema.out[pname];
-        if ($options.types.id.has(C.typeToStringSafe(ptype)))
+        if (_tpLoader.types.id.has(C.typeToStringSafe(ptype)))
             return null;
     }
     return new Ast.Table.Projection(table, ['$event'], table.schema);
@@ -67,7 +70,7 @@ single_stream_projection_String = ('新的' | '新') table:complete_table => {
         if (pname === 'picture_url')
             return null;
         let ptype = table.schema.out[pname];
-        if ($options.types.id.has(C.typeToStringSafe(ptype)))
+        if (_tpLoader.types.id.has(C.typeToStringSafe(ptype)))
             return null;
     }
     return new Ast.Stream.Projection(new Ast.Stream.Monitor(table, null, table.schema), ['$event'], table.schema);
@@ -100,7 +103,7 @@ the_out_param_Array__Any = {
     }
 }
 
-for (let [typestr, type] of $options.types.all) {
+for (let [typestr, type] of _tpLoader.types.all) {
     // don't access booleans or enums out arguments generically, as that rarely makes sense
     // (and when it does, you probably want a macro and maybe and edge trigger)
     if (!type.isEnum && !type.isBoolean) {
@@ -112,12 +115,12 @@ for (let [typestr, type] of $options.types.all) {
             }
         }
 
-        if (!$options.types.id.has(typestr)) {
+        if (!_tpLoader.types.id.has(typestr)) {
             $('projection_' + typestr) = table:complete_table '的' outParam:$('the_out_param_' + typestr) => {
                 const name = outParam.name;
                 if (!table.schema.out[name] || !Type.isAssignable(table.schema.out[name], type))
                     return null;
-                if (name === 'picture_url' && $options.flags.turking)
+                if (name === 'picture_url' && _tpLoader.flags.turking)
                     return null;
                 const newSchema = table.schema.filterArguments((arg) => arg.direction !== Ast.ArgDirection.OUT || arg.name === name);
                 return new Ast.Table.Projection(table, [name], newSchema);
@@ -137,7 +140,7 @@ for (let [typestr, type] of $options.types.all) {
             projection_Any = $('projection_' + typestr);
         }
 
-        if ($options.types.id.has(typestr)) {
+        if (_tpLoader.types.id.has(typestr)) {
             $('single_projection_' + typestr) = table:complete_table => {
                 for (let pname in table.schema.out) {
                     if (table.schema.out[pname].equals(type))
@@ -254,7 +257,7 @@ placeholder_Entity__tt__url = {
     '某个 网址' => Ast.Value.Undefined(true);
 }
 
-for (let [pname, [typestr, pcanonical]] of $options.params.in.values()) {
+for (let [pname, [typestr, pcanonical]] of _tpLoader.params.in.values()) {
     if (typestr.startsWith('Enum__')) {
         // replace $-placeholders with "some X" where X is based on parameter name
 
@@ -293,7 +296,7 @@ for (let [pname, [typestr, pcanonical]] of $options.params.in.values()) {
         action_replace_param_with_table = a:thingpedia_action p:$('single_projection_' + typestr) [-> pname] => C.actionReplaceParamWithTable(a, pname, p);
         action_replace_param_with_stream = a:thingpedia_action p:$('single_stream_projection_' + typestr) [-> pname] => C.actionReplaceParamWithStream(a, pname, p);
 
-        if ($options.types.id.has(typestr)) {
+        if (_tpLoader.types.id.has(typestr)) {
             forward_get_do_command = command:forward_get_do_command '它' [-> pname] => {
                 for (let joinArg in command.table.schema.out) {
                     if (typestr === C.typeToStringSafe(command.table.schema.out[joinArg]))
@@ -341,7 +344,7 @@ for (let [pname, [typestr, pcanonical]] of $options.params.in.values()) {
                     if (pname === 'picture_url')
                         return null;
                     let ptype = command.table.schema.out[pname];
-                    if ($options.types.id.has(C.typeToStringSafe(ptype)))
+                    if (_tpLoader.types.id.has(C.typeToStringSafe(ptype)))
                         return null;
                 }
                 return C.getDoCommand(command, pname, new Ast.Value.Event(null));
@@ -352,7 +355,7 @@ for (let [pname, [typestr, pcanonical]] of $options.params.in.values()) {
                     if (pname === 'picture_url')
                         return null;
                     let ptype = rule.stream.schema.out[pname];
-                    if ($options.types.id.has(C.typeToStringSafe(ptype)))
+                    if (_tpLoader.types.id.has(C.typeToStringSafe(ptype)))
                         return null;
                 }
                 return C.whenDoRule(rule, pname, new Ast.Value.Event(null));
@@ -363,7 +366,7 @@ for (let [pname, [typestr, pcanonical]] of $options.params.in.values()) {
                     if (pname === 'picture_url')
                         return null;
                     let ptype = stream.stream.schema.out[pname];
-                    if ($options.types.id.has(C.typeToStringSafe(ptype)))
+                    if (_tpLoader.types.id.has(C.typeToStringSafe(ptype)))
                         return null;
                 }
                 return C.whenGetStream(stream, pname, new Ast.Value.Event(null));
