+++ languages/thingtalk/zh-cn/timers.genie
