--- lib/common-templates/thingpedia.genie
@@ -1,30 +1,0 @@
-// -*- mode: js; indent-tabs-mode: nil; js-basic-offset: 4 -*-
-//
-// This file is part of ThingTalk
-//
-// Copyright 2017-2018 The Board of Trustees of the Leland Stanford Junior University
-//
-// Author: Giovanni Campagna <gcampagn@cs.stanford.edu>
-//
-// See COPYING for details
-
-// Thingpedia Grammar Rules
-
-thingpedia_query = {}
-thingpedia_get_command = {}
-thingpedia_stream = {}
-thingpedia_action = {}
-thingpedia_program = {}
-thingpedia_who_question = {}
-
-import './common-constants';
-
-// out params nonterminals are automatically generated
-out_param_Any = {}
-out_param_Array__Any = {}
-out_param_npp = {}
-out_param_npi = {}
-out_param_avp = {}
-out_param_pvp = {}
-out_param_npv = {}
-out_param_apv = {}
