+++ lib/genie-compiler/grammar.pegjs
@@ -47,11 +47,11 @@
   }
 }
 
-Grammar = initialComment:$(__) codeBlock:InitialCodeBlock? __ stmt:(Statement __)* {
-  return new Ast.Grammar(initialComment, codeBlock || '', take(stmt, 0));
+Grammar = initialComment:$(__) stmt:(Statement __)* {
+  return new Ast.Grammar(initialComment, take(stmt, 0));
 }
 
-InitialCodeBlock = '{' code:Code '}' { return code; }
+CodeBlock = '{' code:Code '}' { return new Ast.Statement.CodeBlock(code); }
 
 Statement
   = NonTerminalDeclaration
@@ -59,6 +59,7 @@ Statement
   / ForLoop
   / IfStmt
   / ImportStmt
+  / CodeBlock
 
 NonTerminalRef
   = name:Identifier { return new Ast.NonTerminalRef.Identifier(name); }
