+++ lib/genie-compiler/meta_ast.js
@@ -12,9 +12,8 @@
 const { stringEscape } = require('../../lib/escaping');
 
 class Grammar {
-    constructor(comment, initialCode, statements) {
+    constructor(comment, statements) {
         this.comment = comment;
-        this.initialCode = initialCode;
         this.statements = statements;
     }
 
@@ -23,8 +22,7 @@ class Grammar {
 
         buffer += (this.comment);
         buffer += ('"use strict";\n');
-        buffer += (this.initialCode);
-        buffer += `return function($options, $grammar) {`;
+        buffer += `return async function($options, $grammar) {`;
         for (let stmt of this.statements)
             buffer += stmt.codegen('   ');
         buffer += ('   return $grammar;\n');
@@ -37,6 +35,19 @@ exports.Grammar = Grammar;
 
 class Statement {}
 exports.Statement = Statement;
+
+class CodeBlock extends Statement {
+    constructor(code) {
+        super();
+        this.isCodeBlock = true;
+        this.code = code;
+    }
+
+    codegen() {
+        return this.code;
+    }
+}
+Statement.CodeBlock = CodeBlock;
 
 class NonTerminalRef {}
 exports.NonTerminalRef = NonTerminalRef;
@@ -158,7 +169,7 @@ class Import extends Statement {
     }
 
     codegen(prefix = '') {
-        return `${prefix}$grammar = $import(${stringEscape(this.what)})($options, $grammar);\n`;
+        return `${prefix}$grammar = await $import(${stringEscape(this.what)})($options, $grammar);\n`;
     }
 }
 Statement.Import = Import;
